+++ src/Core/Exception/RestrictionException.php
@@ -4,8 +4,8 @@ namespace AppTank\Horus\Core\Exception;
 
 class RestrictionException extends ClientException
 {
-    public function __construct(string $message = "Restriction exception")
+    public function __construct(string $message = "Restriction exception", string|null $code = null, array $context = [])
     {
-        parent::__construct($message);
+        parent::__construct($message, $code, $context);
     }
 }